wave0d.4(#251): move electron/ptyHost/* to daemon (w/ shims)
[REFACTOR-ONLY] No logic changes. File moves + import path updates +
re-export shims + WebContents type-strip. Continues the Wave 0d series
(#955 / #958 / #960). Closes Wave 0d epic #213.

Moves:
- electron/ptyHost/{claudeResolver,cwdResolver,dataFanout,entryFactory,
  index,jsonlResolver,lifecycle,oscTitleSniffer,processKiller}.ts
  -> packages/daemon/src/ptyHost/
- electron/ptyHost/__tests__/*.test.ts (12 specs)
  -> packages/daemon/src/ptyHost/__tests__/*.spec.ts

NodeNext .js extensions added to all relative imports across moved
sources + tests + vi.mock factory paths.

Cross-package import-path updates:
- entryFactory.ts, lifecycle.ts: '../sessionWatcher' -> '../sessionWatcher/index.js'
- jsonlResolver.ts: '../sessionWatcher/projectKey' -> '../sessionWatcher/projectKey.js'
- cwdResolver.ts: '../security/pathGuards' -> '../shared/pathGuards.js' (new
  daemon-local copy, mirrors the #958 precedent for shared/sessionState.ts —
  byte-for-byte identical to electron/security/pathGuards.ts; v0.4 unifies
  both copies via Connect-RPC boundary).

WebContents type-strip (entryFactory.ts):
- Removed `import type { WebContents } from 'electron'` — was the LAST
  electron type-dep in ptyHost. Replaced with a local
  `interface WebContentsLike { isDestroyed(): boolean; send(channel,
  ...args): void }` covering the exact members entryFactory + lifecycle
  + dispatchPtyChunk read on attached senders. Real Electron WebContents
  satisfies this shape structurally.

Daemon deps added (versions copied from root): node-pty 1.1.0,
@xterm/headless ^0.5.5.0, @xterm/addon-serialize ^0.13.0.

Re-export shims at original electron/ paths so out-of-zone consumers
keep compiling unchanged:
- electron/ptyHost/index.ts (consumed by electron/main.ts:62
  killAllPtySessions, electron/notify/bootstrap/installPipeline.ts
  onPtyData)
- electron/ptyHost/oscTitleSniffer.ts (consumed by
  electron/notify/sinks/pipeline.ts)

Other shim entries (claudeResolver, cwdResolver, dataFanout,
entryFactory, jsonlResolver, lifecycle, processKiller) are NOT shimmed
— grep confirmed zero out-of-zone consumers.

Test layout updates:
- vi.mock factory paths updated to .js extensions (vitest matches by
  resolved specifier under NodeNext)
- vi.mock('electron', ...) deletions where no longer needed (entryFactory
  no longer pulls electron)
- Mock interface fields retyped from `ReturnType<typeof vi.fn>` to
  `Mock<(...args) => unknown>` (vitest 4 default Mock type is no longer
  callable via spread; tests now have to typecheck under daemon's
  tsconfig.test.json which includes __tests__, unlike electron's tsconfig
  which excludes them).
- bufferSnapshot.spec.ts: added the now-required `pendingHeadlessWrites`
  + `backpressureWarned` Entry fields to fakeEntry().

electron-load-smoke.test.ts: removed the `ptyHost/dataFanout` entry —
dataFanout no longer lives at electron/ptyHost/dataFanout.ts. Daemon
package owns its own per-module load coverage via
packages/daemon/src/ptyHost/__tests__/*.spec.ts.

Pre-existing daemon dir collision noted: `packages/daemon/src/pty-host/`
(kebab-case, T4.1 child_process pty host) coexists with
`packages/daemon/src/ptyHost/` (camelCase, this PR — matching #958
sessionWatcher convention). Different concerns; intentional.

Push-back on the "delete electron/notify/bootstrap/installPipeline.ts"
half of the task (out of scope here): the file is NOT dead. main.ts:235
still calls `installNotifyPipelineWithProducers` for live wiring of
badge manager (`onNotified` -> badgeManager.incrementSid),
sessionWatcher 'unwatched' -> badgeStore.clearSid (audit #876 H2),
focus/blur producer for decider ctx, OSC sniffer producer wiring, and
`notifyPipelineDispose` is registered on `before-quit` (audit #876
cluster 1.14 / 3.8). 15 vitest cases in the bootstrap __tests__ pin
this contract. The task prompt's claim "this whole pipeline is dead
post-Wave-0c" is incorrect: Wave 0c (#216 / #953) only deleted the IPC
halves of the updater/notify/window registrars; the pipeline + its
producers remain live. Deleting `installPipeline.ts` would break badge
behavior and clean shutdown. Filing back to manager for separate task
scoping; not blocking this move.

Verification:
- pnpm --filter @ccsm/daemon build  GREEN (tsc -p tsconfig.json)
- pnpm --filter @ccsm/daemon typecheck  GREEN (tsc -p tsconfig.json && tsc -p tsconfig.test.json)
- pnpm --filter @ccsm/daemon exec vitest run src/ptyHost  GREEN (12 files, 148 tests)
- pnpm --filter @ccsm/daemon lint  GREEN (0 errors, exit 0)
- npx tsc -p tsconfig.electron.json  GREEN
- npx vitest run tests/electron-load-smoke.test.ts  GREEN (1 file, 14 tests)
- npx vitest run electron/notify/bootstrap/__tests__/installPipeline.test.ts  GREEN (1 file, 15 tests)
- bash tools/lint-no-ipc.sh  PASS
- bash tools/check-v02-shrinking.sh  PASS
