+++ packages/daemon/package.json
@@ -19,7 +19,10 @@
   "dependencies": {
     "@bufbuild/protobuf": "^2.12.0",
     "@connectrpc/connect": "^2.1.1",
-    "better-sqlite3": "^12.9.0"
+    "@xterm/addon-serialize": "^0.13.0",
+    "@xterm/headless": "^5.5.0",
+    "better-sqlite3": "^12.9.0",
+    "node-pty": "1.1.0"
   },
   "devDependencies": {
     "@bufbuild/protobuf": "^2.12.0",
