+++ packages/daemon/src/ptyHost/claudeResolver.ts
