+++ packages/daemon/src/ptyHost/dataFanout.ts
